index.html
added index.html to the repository after setting up the barebones structure for the recipe website. Included are pictures and hyperlink references to each recipe

recipes/kfries.html

added kfries.html to the repository. Set up the basic structure for the recipe web page and an image

recipes/lambasada.html

added lambasada.html to the repository. Set up the basic structure for the recipe webpage and an image

recipes/mangochicken.html

added mangochicken.html to the repository. Set up the basic structure for the recipe webpage and an image